// src/model/interfaces/cartaAceptacion/datosCartaAceptacion.ts
export default interface DatosCartaAceptacion{
    nombreAnio: string
    ciudadFecha: string
    nombreReferente: string
    empresaReferente: string
    cargoReferente: string
    curso: string
    nombreEstudiante: string
    nivelEstudiante: string
    carrera: string
    facultad: string
    nombreAdministrativo: string
    cargoAdministrativo: string
    urlFirma: string
}